refactor: profile file names to dashboard
