+++ src/constants/routes.js
@@ -1,7 +1,7 @@
 export const HOME = "/home";
 export const MYCOURSE = "/my-course";
 export const WISHLIST = "/wishlist";
-export const PROFILE = "/profile";
+export const PROFILE = "/dashboard";
 export const SIGNUP = "/signup";
 export const COURSE_DETAILS = "/course/:id";
 export const LESSONS_PLAYER = "/course/:courseName/lecture/:videoSectionId:videoSectionName";
